Save system ver2
You can now load the caravan that you built (and saved) in the Journey Stage
