+++ Assets/Prefabs/JourneyStage/SWagon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dd3452bb3b2e5914eb98fbcd07437753
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
