+++ Assets/Prefabs/Wagon/Crate.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fc209a27ac448e74a86e388cdad8b184, type: 3}
   m_Name: Crate
   m_EditorClassIdentifier: 
-  storage_name: Crate
+  id: crate
   size: {x: 2, y: 2, z: 2}
   storage_type: 0
   inventorySize: {x: 5, y: 5}
