+++ Assets/Prefabs/Wagon/LargeCage.asset
@@ -12,7 +12,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fc209a27ac448e74a86e388cdad8b184, type: 3}
   m_Name: LargeCage
   m_EditorClassIdentifier: 
-  storage_name: LargeCage
+  id: large_cage
   size: {x: 5, y: 4, z: 6}
   storage_type: 0
   inventorySize: {x: 5, y: 5}
