+++ Assets/Scripts/CaravanCustomization/Upgrade_CaravanManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6897dbc7c0a754c45a138953087f94c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
