+++ Assets/Scripts/CaravanCustomization/Wagon.cs
@@ -32,15 +32,25 @@ public class Wagon : MonoBehaviour
     private void PlaceModule(StorageModule selectedModule)
     {
         float yRotation = selectedModule.rotation;
+
+        // Get the world position based on grid rotation and position
         Vector3 placementPosition = grid.GridToAdjustedWorldPosition(selectedModule.currentPosition, selectedModule.Size);
-        // Instantiate the module and set its properties
+
+        // Combine grid rotation with the module's intended rotation
+        Quaternion gridRotation = grid.gridRotation;
+        Quaternion moduleRotation = Quaternion.Euler(0, yRotation, 0);
+
+        // Instantiate the module with combined rotation
         GameObject placedModule = Instantiate(
             selectedModule.moduleData.graphics,
             placementPosition,
-            Quaternion.Euler(0, yRotation, 0),
+            gridRotation * moduleRotation, // Apply grid rotation as base, then module's relative rotation
             transform
         );
+
+        // Additional settings
         selectedModule.objectRef = placedModule;
         placedModule.layer = 7;
     }
+
 }
