+++ Assets/Scripts/JourneyStage/Journey_CaravanManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07b5b9ff4a7e5ed4da78140e42c2ca9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
