+++ Assets/Scripts/SaveSystem/ICaravanDataReceiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f7e3cf236385e0e46813d57676ceb5a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
